requirements.txt
# base requirements

Django==3.2.13
djangorestframework==3.11.2
Unipath==1.1
colorlog==2.6.0
ConcurrentLogHandler==0.9.1
#MySQL-python==1.2.5
#mysqlclient==1.3.12
PyMySQL==1.0.2
django_debug_toolbar==3.2.1
django-debug-toolbar-request-history==0.1.3
django-extensions==2.0.7
django-ratelimit==2.0.0
Pillow==9.0.1
django-precise-bbcode==1.2.14
celery==5.2.3
requests==2.20.0
html2bbcode==2.3.3
wiki==0.7.9
pyyaml==5.4.1
user-agents==2.1
gunicorn==20.1.0
